# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SchoolHealth AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visits:
    get:
      operationId: listVisits
      tags: [visits]
      summary: List nurse visits with optional filters
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: grade
          in: query
          schema:
            type: string
        - name: classroom
          in: query
          schema:
            type: string
        - name: symptom
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of visits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitListResponse"
    post:
      operationId: createVisit
      tags: [visits]
      summary: Log a new nurse visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVisitBody"
      responses:
        "201":
          description: Visit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}:
    get:
      operationId: getVisit
      tags: [visits]
      summary: Get a single visit by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: grade
          in: query
          schema:
            type: string
        - name: classroom
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentListResponse"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a new student record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student with their health history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details with visit history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetail"
...
components:
  schemas:
...
    Visit:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        grade:
          type: string
        classroom:
          type: string
        symptoms:
          type: array
          items:
            type: string
        temperature:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        actionTaken:
          type: string
          enum: [sent_home, returned_to_class, called_parent, referred_to_doctor, monitored]
        visitDate:
          type: string
          format: date-time
        loggedBy:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - studentName
        - grade
        - classroom
        - symptoms
        - actionTaken
        - visitDate
        - loggedBy
        - createdAt
    VisitListResponse:
      type: object
      properties:
        visits:
          type: array
          items:
            $ref: "#/components/schemas/Visit"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - visits
        - total
        - limit
        - offset
    CreateVisitBody:
      type: object
      properties:
        studentId:
          type: integer
          nullable: true
        anonymousId:
          type: string
          nullable: true
        grade:
          type: string
        classroom:
          type: string
        symptoms:
          type: array
          items:
            type: string
          minItems: 1
        temperature:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        actionTaken:
          type: string
          enum: [sent_home, returned_to_class, called_parent, referred_to_doctor, monitored]
      required:
        - grade
        - classroom
        - symptoms
        - actionTaken
    Student:
      type: object
      properties:
        id:
          type: integer
        studentCode:
          type: string
        name:
          type: string
        grade:
          type: string
        classroom:
          type: string
        dateOfBirth:
          type: string
          nullable: true
        parentEmail:
          type: string
          nullable: true
        parentPhone:
          type: string
          nullable: true
        chronicConditions:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
        visitCount:
          type: integer
        lastVisitDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studentCode
        - name
        - grade
        - classroom
        - chronicConditions
        - allergies
        - visitCount
        - createdAt
    StudentListResponse:
      type: object
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
      required:
        - students
        - total
    CreateStudentBody:
      type: object
      properties:
        name:
          type: string
        grade:
          type: string
        classroom:
          type: string
        dateOfBirth:
          type: string
          nullable: true
        parentEmail:
          type: string
          nullable: true
        parentPhone:
          type: string
          nullable: true
        chronicConditions:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
      required:
        - name
        - grade
        - classroom
    StudentDetail:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
        recentVisits:
          type: array
          items:
            $ref: "#/components/schemas/Visit"
        riskScore:
          type: number
          description: Predicted risk score 0-100
      required:
        - student
        - recentVisits
        - riskScore
...